Set TMC_UART::getTime to return 32bit value because size_t is 16bit on AVR

+++ src/source/TMC_HAL/TMC_HAL_AVR.cpp
@@ -115,7 +115,7 @@ int TMC_UART::available() {
 }
 
 __attribute__((weak))
-size_t TMC_UART::getTime() const {
+uint32_t TMC_UART::getTime() const {
     return millis();
 }
